Adds refactor to night_writer class to move all methods and instance variables inside translator as parent class
